feat(#1529): delete NexusServiceClient + deprecate HTTP RPC endpoint
- Delete src/nexus/cli/client.py (NexusServiceClient, ~306 lines)
- Delete get_service_client() from utils.py
- Move NexusAPIError to cli/clients/base.py (used by domain clients)
- Update service_command.py: client_class is now required
- Add Deprecation/Sunset/X-Migration-Guide headers to HTTP RPC endpoint
- Delete test_service_client.py and test_client.py (~805 lines)
- Remove mock_service_client/patch_service_client fixtures from conftest
- Skip E2E tests pending gRPC migration
- Update all NexusAPIError imports to cli.clients.base
