+++ tests/unit/cli/conftest.py
@@ -10,7 +10,7 @@ import pytest
 import yaml
 from click.testing import CliRunner
 
-from nexus.cli.client import NexusAPIError
+from nexus.cli.clients.base import NexusAPIError
 from nexus.cli.config import NexusCliConfig, ProfileEntry
 
 
@@ -128,22 +128,6 @@ def patch_compose_runner_status(mock_compose_runner: MagicMock):
 # ---------------------------------------------------------------------------
 # Service command fixtures (Issue #2812)
 # ---------------------------------------------------------------------------
-
-
-@pytest.fixture()
-def mock_service_client():
-    """Mock NexusServiceClient for testing service CLI commands."""
-    client = MagicMock()
-    client.__enter__ = MagicMock(return_value=client)
-    client.__exit__ = MagicMock(return_value=False)
-    return client
-
-
-@pytest.fixture()
-def patch_service_client(mock_service_client):
-    """Patch get_service_client to return mock. Yields the mock client."""
-    with patch("nexus.cli.service_command.get_service_client", return_value=mock_service_client):
-        yield mock_service_client
 
 
 def make_api_error(status_code: int = 500, detail: str = "Internal Server Error") -> NexusAPIError:
